changed Router folder (one folder for each page)
